Make node_modules_paths available to resolver core logic, rather than just default_resolver.
